requirements.txt
google-api-python-client==2.176.0
google-cloud-bigquery==3.35.0
grpcio==1.73.1
google-auth-oauthlib==1.2.2
confuse==2.0.1
slack-sdk==3.36.0
requests==2.32.4
functions-framework==3.8.3
google-cloud-secret-manager==2.24.0
